refactor(lifecycle): extract kukeond ready-check helpers and relocate package to cmd/kuke/internal/lifecycle
